[UPD] change reset logic from register <-> RegistrationForm componenent
